moved addroot functionality into texmagician component
