+++ package.json
@@ -214,7 +214,7 @@
       },
       {
         "command": "latex-workshop.addroot",
-        "title": "Insert root document",
+        "title": "Insert %!TeX root magic command",
         "category": "LaTeX Workshop"
       },
       {
